+++ tests/integration/unknown_values/program/index.ts
@@ -1,0 +1,11 @@
+// Copyright 2016-2020, Pulumi Corporation.  All rights reserved.
+
+import * as random from "@pulumi/random";
+
+const str = new random.RandomString("str", {
+    length: 10,
+});
+
+const pet = new random.RandomPet("pet", {
+    prefix: str.result,
+});